+++ CHANGELOG.md
@@ -24,6 +24,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - [#2315](https://github.com/clj-kondo/clj-kondo/issues/2315): fix regression with unused JavaScript namespace
 - [#2302](https://github.com/clj-kondo/clj-kondo/issues/2302): New linter: `:equals-expected-position` to enforce expected value to be in first (or last) position. See [docs](https://github.com/clj-kondo/clj-kondo/blob/master/doc/linters.md)
 - [#2304](https://github.com/clj-kondo/clj-kondo/issues/2304): Report unused value in `defn` body
+- [#2227](https://github.com/clj-kondo/clj-kondo/issues/2227): allow `:flds` to be used in keys destructuring for ClojureDart
 
 ## 2024.03.13